requirements.txt
flask
google-generativeai
langchain
langchain-google-genai
faiss-cpu
python-dotenv